% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUnifrac.R
\name{scUnifrac}
\alias{scUnifrac}
\title{scUnifrac}
\usage{
scUnifrac(data1, sampleName1 = "S1", data2, sampleName2 = "S2",
  ref.expr = NULL, genenum = 500, ncluster = 10, nDim = 4,
  normalize = T, report = T, outputFile = "scUnifrac_report.html")
}
\arguments{
\item{data1}{matrix; the data matrix of the first dataset, row is the gene symbol, the column is the cell id}

\item{sampleName1}{a character string giving the name of the first dataset; (default: "S1")}

\item{data2}{matrix; the data matrix of the second dataset, row is the gene symbol, the column is the cell id; data1 and data1 should have the same gene symbols}

\item{sampleName2}{a character string giving the name of the second dataset; (default: "S2")}

\item{ref.expr}{matrix; the data matrix of the reference cell atlas; the ref.expr is used to predict the cell types of single cell from data1 and data2 (default: NULL)}

\item{genenum}{integer; Number of highly variable genes to build the cell population structure (default: 500)}

\item{ncluster}{integer; Number of clusters to divide cells  (default: 10)}

\item{nDim}{integer; Number of PCA dimensions to build the cell population structure  (default: 4)}

\item{normalize}{logical; Indicate whether normalize data1 and data2 (default: TRUE, normalize to the total count and log2 transform)}

\item{report}{logical;  Indicate whether to generate a report (default: TRUE); if set to FALSE, scUnifrac will not generate the report but calculate the distance and the pvalue; set to FALSE, when users have more than two samples to compare and only want to calculate the pairwise distance}

\item{outputFile}{a character string giving the name of the report;(default: "scUnifrac_report.html")}
}
\value{
List with the following elements:
\item{distance}{The distance of cell population diversity between two single-cell RNA-seq datasets}
\item{p-value}{The statistical signficance of the distance}
}
\description{
Quantify cell population diversity between two single cell RNA-seq datasets
}
\examples{
#library(scUnifrac)  
##load the two example datasets 
#load(system.file("extdata", "colon1.Rdata", package = "scUnifrac"))
#load(system.file("extdata", "pan1.Rdata", package = "scUnifrac"))

#result<-scUnifrac( data1=colon1, data2=pan1)  #this function will return distance and pvalue between data1 and data2 and will generate a report in the work directory
#result
##load the mouse cell altas from Han et al., 2018, Cell 172, 1091–1107. The atlas is used as a reference to predict cell types of data1 and data2
#load(system.file("extdata", "ref.expr.Rdata", package = "scUnifrac"))
#result<-scUnifrac( data1=colon1, data2=pan1,ref.expr=ref.expr, outputFile="scUnifrac_report.html") #the report includes the predicted cell types of each cell in each differential subpopulation between data1 and data2

##test two samples with similar populations
#ind<-sample(c(1:ncol(colon1)), ncol(colon1)/2)
#result<-scUnifrac(data1=colon1[,ind], data2=colon1[,-ind],ref.expr=ref.expr, outputFile="scUnifrac_report.html")

}
